Moving all testing models to their own file; preparing to write a more complete Meta test.
