+++ response_test.go
@@ -3,109 +3,11 @@ package jsonapi
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
 	"reflect"
 	"sort"
 	"testing"
 	"time"
 )
-
-type Blog struct {
-	ID            int       `jsonapi:"primary,blogs"`
-	ClientID      string    `jsonapi:"client-id"`
-	Title         string    `jsonapi:"attr,title"`
-	Posts         []*Post   `jsonapi:"relation,posts"`
-	CurrentPost   *Post     `jsonapi:"relation,current_post"`
-	CurrentPostID int       `jsonapi:"attr,current_post_id"`
-	CreatedAt     time.Time `jsonapi:"attr,created_at"`
-	ViewCount     int       `jsonapi:"attr,view_count"`
-}
-
-func (b *Blog) JSONAPILinks() *Links {
-	return &Links{
-		"self": fmt.Sprintf("https://example.com/api/blogs/%d", b.ID),
-		"comments": Link{
-			Href: fmt.Sprintf("https://example.com/api/blogs/%d/comments", b.ID),
-			Meta: Meta{
-				"counts": map[string]uint{
-					"likes":    4,
-					"comments": 20,
-				},
-			},
-		},
-	}
-}
-
-func (b *Blog) JSONAPIRelationshipLinks(relation string) *Links {
-	if relation == "posts" {
-		return &Links{
-			"related": Link{
-				Href: fmt.Sprintf("https://example.com/api/blogs/%d/posts", b.ID),
-				Meta: Meta{
-					"count": len(b.Posts),
-				},
-			},
-		}
-	}
-	if relation == "current_post" {
-		return &Links{
-			"self": fmt.Sprintf("https://example.com/api/posts/%s", "3"),
-			"related": Link{
-				Href: fmt.Sprintf("https://example.com/api/blogs/%d/current_post", b.ID),
-			},
-		}
-	}
-	return nil
-}
-
-func (b *Blog) JSONAPIMeta() *Meta {
-	return &Meta{
-		"detail": "extra details regarding the blog",
-	}
-}
-
-func (b *Blog) JSONAPIRelationshipMeta(relation string) *Meta {
-	if relation == "posts" {
-		return &Meta{
-			"detail": "extra posts detail",
-		}
-	}
-	if relation == "current_post" {
-		return &Meta{
-			"detail": "extra current_post detail",
-		}
-	}
-	return nil
-}
-
-type Post struct {
-	Blog
-	ID            uint64     `jsonapi:"primary,posts"`
-	BlogID        int        `jsonapi:"attr,blog_id"`
-	ClientID      string     `jsonapi:"client-id"`
-	Title         string     `jsonapi:"attr,title"`
-	Body          string     `jsonapi:"attr,body"`
-	Comments      []*Comment `jsonapi:"relation,comments"`
-	LatestComment *Comment   `jsonapi:"relation,latest_comment"`
-}
-
-type Comment struct {
-	ID       int    `jsonapi:"primary,comments"`
-	ClientID string `jsonapi:"client-id"`
-	PostID   int    `jsonapi:"attr,post_id"`
-	Body     string `jsonapi:"attr,body"`
-}
-
-type Book struct {
-	ID          uint64  `jsonapi:"primary,books"`
-	Author      string  `jsonapi:"attr,author"`
-	ISBN        string  `jsonapi:"attr,isbn"`
-	Title       string  `jsonapi:"attr,title,omitempty"`
-	Description *string `jsonapi:"attr,description"`
-	Pages       *uint   `jsonapi:"attr,pages,omitempty"`
-	PublishedAt time.Time
-	Tags        []string `jsonapi:"attr,tags"`
-}
 
 func TestMarshall_attrStringSlice(t *testing.T) {
 	tags := []string{"fiction", "sale"}
